First PR for msal-node, moved around some code to fit the new project scaffolding
